Update Progression.txt
Changes to somethings that were marked incorrectly. Changes were done to non-Calamity Terraria stuff. Some Post-<event flag> stuff was marked as filler or useful, but should be marked as progression and vice versa. Reward stuff was changed in sense that if something would help with a check it was marked as useful, as all rewards are obtainable regularly in game, and if not marked as filler. No matter what "Reward: Star Cloak" is incorrectly marked as a trap.