Final Fantasy - Reclassify mcguffins
I reclassified Shard and the Orbs as mcguffins, because they are required for goal but not for any multiworld checks (the chests in the final dungeon are never multiworld locations).  On the standard goal, you need all four orbs to fight the final boss.  On the shard hunt goal, you need some amount of shards to fight the final boss.  This change doesn't even matter for orbs, because they're always local items, not multiworld items.  Shards can be multiworld items.