Aquaria - Classified new duplicate items and reclassified Sun Key (#32)
* Aquaria - Classified new duplicate items with different capitalization

The bot has added a few new capitalization versions of item names since Palex did his major cleanup.  This is just setting the classification of the new names to the same thing as the other capitalization version.

* Aquaria - Reclassify Sun Key as filler

Sun Key is not actually useful at all, even though it's currently classified as useful.  I asked the developer about this, and he said he would reclassify it as filler in the next apworld update.  This applies that change early.

https://discord.com/channels/731205301247803413/1257452465272389687/1355231947713941564