SML2 - Updated the whole list (#41)
Classified many unknown items.
Reclassified some midway bells and cancel auto scrolls to match the apworld.
Reclassified some of the named coins as mcguffins (they were inconsistently classified).
Added a few item names from an older version of the apworld since most of the names from that apworld were already in the list.

There are a few ambiguous items in the list for Auto Scroll and Cancel Auto Scroll.  I believe there are typos in the apworld code causing it to generate items that were intended to be skipped, and possibly not generate items that it intended to include.  The code is still clear on which Cancel Auto Scrolls are supposed to be useful, so I followed that intention.  I think the Auto Scroll and Cancel Auto Scroll for Mario's Castle were intended to not exist, so they don't have defined classifications.  I classified them based on what I believe they should be if they do exist, based on my game knowledge.