Earthbound - Update all classifications (#34)
I went through the entire list and updated the classifications to match the apworld.  There is definitely a lot of gray area in how useful something must be to be considered useful, so it makes since that users would have classified things inconsistently.

The only two oddballs in the list are ATM Card and Sound Stone, since you start with both of them; they're not in the multiworld.  I called the ATM Card progression and Sound Stone mcguffin because if the apworld ever changed so that you didn't start with them, that's what I'd expect them to be.  (ATM Card progression for shopsanity checks; Sound Stone unlocking no checks, just goal)