DKC3 - Reclassify mcguffins (#33)
* DKC3 - Reclassify mcguffins

Banana Bird is definitely a mcguffin; they unlock no checks but are needed to goal if the goal is Banana Bird Hunt (15 by default).
Bonus Coins are needed to unlock levels (with checks) in the last world (15/level by default), and all of those levels need to be unlocked in order to goal if the goal is Knautilus.
Krematoa Cogs are only needed to reach Knautilus for that goal; 5 are needed.
DK Coins are used to get the Gyrocopter, which is progression and not needed for goal.  However, you need 10-41 of them to unlock it (30 by default).  That seems like it matches this bot's definition of a mcguffin.

* DKC3 - Reclassify Donkey Kong as mcguffin

On review, I'd call Donkey Kong a mcguffin as well.  It's just the "Victory" item.