Changed some Hollow Knight Charms and all Whispering Root from progression to useful (#59)
Hollow Knight Charms are marked as progression in the apworld because they unlock Salubra checks, though really only the Baldur Killer ones unlock progression. Marking most of them as useful instead of progression helps with keeping down the messaging when set to the proper filtering level.

In similar spirit, the Whispering Roots give such minimal amounts of essence, that their usefulness for progression in the essence shop is neglible, so marking them as useful instead of progression helps keep down the messaging when set to progression only filter.