Moved World Sense, Dream Wielder to useful, marked Grubs and Eggs as mcguffin
Missed that Dream Wielder Charm was marked as progression still, since it was hiding among the Dream Nails

World Sense is a bit weird to be marked as progression because all it does is show world completion % and Hunter's Journal Enemies killed/required to be killed, really it should be filler, but there is some use for journal rando maybe? I do not know if the Rando defaults to showing the Enemies required without it. 

Marked both grubs and rancid eggs as mcguffin, since they dont do anything on their own, other than get you checks/money from Grubfather and Rancid Egg Shop checks(if enabled since default is disabled for Egg Shop)